backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pydantic>=2.0.0
pydantic-settings
pytest
python-dotenv
transformers
torch
pdfplumber

python-multipart
